--- src/components/Navbar.jsx
@@ -1,31 +1,0 @@
-import React from 'react'
-import NavBar from "../components/css/NavBar.css"
-
-const Navbar = () => {
-  return (
-    <>
-     <nav class="navbar navbar-expand-lg bg-body-tertiary">
-  <div class="container-fluid">
-    <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-      <span class="navbar-toggler-icon"></span>
-    </button>
-    <div class="collapse navbar-collapse" id="navbarNav">
-      <ul class="navbar-nav">
-        <li class="nav-item">
-          <a class="nav-link active" aria-current="page" href="#">Home</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" href="#">Login</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" href="#">SingUp</a>
-        </li>
-      </ul>
-    </div>
-  </div>
-</nav>
-    </>
-  )
-}
-
-export default Navbar
