--- src/components/css/NavBar.css
@@ -1,6 +1,0 @@
-.navbar-nav{
-    display: inline-flex;
-}
-.nav-item{
-    margin: 2px;
-}
