--- src/components/pages/Dashboard.jsx
@@ -1,11 +1,0 @@
-import Navbar from "../Navbar"
-
-const Dashboard = () => {
-  return (
-    <div className="centered">
-      <Navbar />
-    </div>
-  )
-}
-
-export default Dashboard
