--- src/components/pages/Login.jsx
@@ -1,7 +1,0 @@
-const Login = () => {
-  return (
-    <div className="centered">Login</div>
-  )
-}
-
-export default Login
