--- src/components/pages/Register.jsx
@@ -1,7 +1,0 @@
-const Register = () => {
-  return (
-    <div className="centered">Register</div>
-  )
-}
-
-export default Register
