+++ src/pages/Dashboard.jsx
@@ -1,11 +1,11 @@
 import React from 'react'
 
-const App = () => {
+const Dashboard = () => {
   return (
     <div>
-      App React
+      
     </div>
   )
 }
 
-export default App
+export default Dashboard
