+++ src/pages/Login.jsx
@@ -1,11 +1,11 @@
 import React from 'react'
 
-const App = () => {
+const Login = () => {
   return (
     <div>
-      App React
+      
     </div>
   )
 }
 
-export default App
+export default Login
